Fixes from remarks 2.1

+++ aleksandrov.igor/P4/latinRemove.cpp
@@ -1,10 +1,11 @@
 #include "latinRemove.h"
 #include <cctype>
+#include <cstddef>
 
 char* aleksandrov::latinRemove(char* str)
 {
-  int j = 0;
-  for (int i = 0; str[i] != '\0'; ++i)
+  size_t j = 0;
+  for (size_t i = 0; str[i] != '\0'; ++i)
   {
     if (!std::isalpha(str[i]))
     {
